Rename modules for preferred 'retention strategy' naming.
